+++ src/features/reader/overlay/navigation/desktop/components/ReaderNavBarDesktopActions.tsx
@@ -26,8 +26,11 @@ import {
     getReaderPagesStore,
     useReaderChaptersStore,
     useReaderPagesStore,
+    useReaderStore,
 } from '@/features/reader/stores/ReaderStore.ts';
 import type { ChapterDownloadInfo, ChapterIdInfo } from '@/features/chapter/Chapter.types.ts';
+import { defaultPromiseErrorHandler } from '@/lib/DefaultPromiseErrorHandler.ts';
+import ShareIcon from '@mui/icons-material/Share';
 
 const DownloadButton = ({ id = -1, isDownloaded }: ChapterIdInfo & ChapterDownloadInfo) => {
     const { t } = useLingui();
@@ -64,8 +67,10 @@ const DownloadButton = ({ id = -1, isDownloaded }: ChapterIdInfo & ChapterDownlo
 };
 
 export const ReaderNavBarDesktopActions = memo(() => {
-    const { id, isDownloaded, isBookmarked, realUrl } = useReaderChaptersStore((state) => ({
+    const mangaTitle = useReaderStore((state) => state.manga?.title);
+    const { id, name, isDownloaded, isBookmarked, realUrl } = useReaderChaptersStore((state) => ({
         id: state.currentChapter?.id ?? FALLBACK_CHAPTER.id,
+        name: state.currentChapter?.name ?? FALLBACK_CHAPTER.name,
         isDownloaded: state.currentChapter?.isDownloaded ?? FALLBACK_CHAPTER.isDownloaded,
         isBookmarked: state.currentChapter?.isBookmarked ?? FALLBACK_CHAPTER.isBookmarked,
         realUrl: state.currentChapter?.realUrl ?? FALLBACK_CHAPTER.realUrl,
@@ -120,6 +125,24 @@ export const ReaderNavBarDesktopActions = memo(() => {
                     <IconWebView />
                 </IconButton>
             </CustomTooltip>
+            {'share' in navigator && (
+                <CustomTooltip title={t`Share`} disabled={!realUrl}>
+                    <IconButton
+                        color="inherit"
+                        disabled={!realUrl}
+                        onClick={() =>
+                            navigator
+                                .share({
+                                    title: `${mangaTitle} - ${name}`,
+                                    url: realUrl ?? undefined,
+                                })
+                                .catch(defaultPromiseErrorHandler('MangaToolbarMenu::share'))
+                        }
+                    >
+                        <ShareIcon />
+                    </IconButton>
+                </CustomTooltip>
+            )}
         </Stack>
     );
 });